requirements.txt
flask
opencv-python>=4.8
ultralytics>=8.0
pymongo>=4.6
python-dotenv>=1.0
gunicorn>=21.0         
numpy
python-dotenv
twilio
easyocr
yt-dlp
python-dotenv
certifi
